refactor: 하이라이트 타입을 하위 모듈로 분리
- highlight.ts 단일 파일을 types/highlight/ 디렉토리로 분할
- input-logs, segments, insight-tags, upload, transformers 모듈화
- API 응답 타입(snake_case)과 클라이언트 타입(camelCase) 분리
